changes route from /map to /compare and renames Map.js to Comparison.js
