+++ src/components/pages/comparison/Comparison.js
